src/backend/requirements.txt
fastapi
uvicorn[standard]
requests
pyjwt
python-multipart
pysqlite3